Expose tags to configuration providers
