+++ cuppa/src/main/java/org/forgerock/cuppa/expression/ContainsCondition.java
@@ -14,7 +14,28 @@
  * limitations under the License.
  */
 
+package org.forgerock.cuppa.expression;
+
+import java.util.Collection;
+
 /**
- * Classes related to the expression group filter.
+ * A condition that checks if a tag is contains in a collection of tags.
  */
-package org.forgerock.cuppa.internal.filters.expression;
+public class ContainsCondition implements Condition {
+
+    private String tag;
+
+    /**
+     * Constructor.
+     *
+     * @param tag A group/tag we want to search for.
+     */
+    ContainsCondition(String tag) {
+        this.tag = tag;
+    }
+
+    @Override
+    public boolean shouldRun(Collection<String> tags) {
+        return tags.contains(tag);
+    }
+}
